Make pip installation work: pip install -e .
The reason it was not working is that there is the default_mountpoint
directory too, and then "import csync" would fail (from the executable
created via pyproject.toml).

But if csync.py is in its src/ directory, python does not get confused
anymore.

For a more interesting/complex/nuanced way of creating executables
with pip and at the same time letting the original files be
executables, see the pyproject.toml and dir and file structure in
http://gitlab.com/jordibc/mapelia .